changed depth_control/state and /setpoint to Int16

+++ Mission Control/competition_tasks/src/bouyTask.py
@@ -54,8 +54,8 @@ class TrackObjectState(smach.State):
 		self.yaw_publisher = rospy.Publisher('/yaw_control/setpoint', Float64, queue_size=10) # desired orientation
 
 		self.depth_current = 0 # in inches
-		rospy.Subscriber('/depth_control/state', Float64, self.depth_callback)
-		self.depth_publisher = rospy.Publisher('/depth_control/setpoint', Float64, queue_size=10)
+		rospy.Subscriber('/depth_control/state', Int16, self.depth_callback)
+		self.depth_publisher = rospy.Publisher('/depth_control/setpoint', Int16, queue_size=10)
 
 		self.forward_thrust_publisher = rospy.Publisher('/yaw_pwm', Int16, queue_size=10)
 		self.forward_thrust = 0
@@ -171,8 +171,8 @@ class ChangeDepthState(smach.State):
 		self.threshold = threshold
 
 		self.depth_current = 0 # in inches
-		rospy.Subscriber('/depth_control/state', Float64, self.depth_callback)
-		self.depth_publisher = rospy.Publisher('/depth_control/setpoint', Float64, queue_size=10)
+		rospy.Subscriber('/depth_control/state', Int16, self.depth_callback)
+		self.depth_publisher = rospy.Publisher('/depth_control/setpoint', Int16, queue_size=10)
 
 		self.has_reset = False
 		self.reset_subscriber = rospy.Subscriber('/reset', Bool, self.reset_callback)
